requirements.txt
requests==2.26.0
PyGithub==1.55
python-dotenv==0.19.1
